decouple global_user from global_config
