+++ src/settings/mod.rs
@@ -1,7 +1,9 @@
 pub mod binding;
 mod environment;
+mod global_config;
 pub mod global_user;
 pub mod metadata;
 pub mod toml;
 
 pub use environment::{Environment, QueryEnvironment};
+pub use global_config::{get_global_config_path, get_wrangler_home_dir, DEFAULT_CONFIG_FILE_NAME};
